# Check for buggy preadv2