Closes: #3803

cc @br0adcast @wmanley