@Darksonn 2021-06-02
@wmanley I don't really understand the Linux kernel version numbers here. The Linux git repository only has tags for `v5.10` and `v5.11`, and not for things like `v5.10.17` like the version from the original issue.