@wmanley 2021-06-02
> @wmanley I don't really understand the Linux kernel version numbers here. The Linux git repository only has tags for `v5.10` and `v5.11`, and not for things like `v5.10.17` like the version from the original issue.

Here's the git repo with the stable kernel versions: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git .  This is where the point releases are tagged.  I've tested both the mainline and the stable kernel tags.