@Darksonn 2021-06-02
> I've tested both the mainline and the stable kernel tags.

I got the impression that you just did a bisection? Which tags did you try?