@wmanley 2021-06-02
> > I've tested both the mainline and the stable kernel tags.
> 
> I got the impression that you just did a bisection? Which tags did you try?

I did both.

I did a bisection on mainline to identify the specific commits where it was introduced and fixed. Having discovered that it was fixed on mainline I separately and additionally ran the test on the latest stable releases to see if the fix had been cherry-picked to any stable branch.  The results of the stable branch checking is as follows:

```
5.6.19: OK
5.7.19: OK
5.8.18: OK
5.9.16: BAD
5.10.38: BAD
5.11.22: OK
```

My conclusion is that the kernel developers aren't/weren't even aware that the bug existed, so the fix for it hasn't been backported.

Bisecting to find first bad commit: https://github.com/tokio-rs/tokio/issues/3803#issuecomment-845355092
Bisecting to find first fixed commit: https://github.com/tokio-rs/tokio/issues/3803#issuecomment-845414528
Testing stable branches: https://github.com/tokio-rs/tokio/issues/3803#issuecomment-845879810