@carllerche 2021-06-02
Before merging this, I think that we need to have a plan to include testing different kernel versions as part of CI. we don't want to accidentally introduce regressions if we change this code.

Granted, testing different kernel versions is not trivial.