@blasrodri 2021-07-05
> Before merging this, I think that we need to have a plan to include testing different kernel versions as part of CI. we don't want to accidentally introduce regressions if we change this code.
> 
> Granted, testing different kernel versions is not trivial.

I have not had the time to check in detail. But this project seems like an alternative https://github.com/linuxkit/linuxkit