@wmanley 2021-07-12
> Before merging this, I think that we need to have a plan to include testing different kernel versions as part of CI. we don't want to accidentally introduce regressions if we change this code.

I'd be happy to have a crack at this, adapting the kernel testing I'd done in #3803 to CI.  So we'd have a few pre-built kernels and run qemu in CI to see if the bug is present or not on each. Ideally we'd have a way of running against the latest RC kernel too to catch future regressions before they happen.

But first I'd like to know - is it worthwhile? Does anyone have a real-world benchmark that this helps with, or is there some other reason to want this?