@Darksonn 2021-07-12
A benchmark? It's mostly to make sure we catch regressions if we break something on some kernel versions only.