@wmanley 2021-07-12
> A benchmark? It's mostly to make sure we catch regressions if we break something on some kernel versions only.

Not a benchmark for testing, but for #3518. The value in the kernel testing is to see if we can un-revert #3518, so the question is of whether #3518 is of sufficient value to bother. I wrote #3518 for fun rather than to satisfy a particular need that I have - I'm not actually a tokio user. I'm happy to have a go at test-infra, but not for it's own sake - only if it's actually providing someone with value. So it comes down to - is this something that will actually help people?