@Darksonn 2021-07-12
Generally, file reading involves a whole bunch of cross-thread communication with the current implementation, so it would surprise me if it doesn't help.

I tried running the benchmarks in `benches/fs.rs` with and without the change.
```
Without the change
test async_read_buf      ... bench:   6,714,074 ns/iter (+/- 1,063,411)
test async_read_codec    ... bench:   6,727,703 ns/iter (+/- 893,984)
```
```
With the change
test async_read_buf      ... bench:   6,363,540 ns/iter (+/- 807,035)
test async_read_codec    ... bench:   6,454,059 ns/iter (+/- 857,237)
```
Though it's just reading from `/dev/zero`. I don't know how representative that is.